+++ linkerd/server-policy/src/authz.rs
@@ -50,7 +50,7 @@ impl Suffix {
 #[cfg(feature = "proto")]
 pub mod proto {
     use super::*;
-    use crate::{authz::Network, proto::InvalidMeta};
+    use crate::{authz::Network, meta::proto::InvalidMeta};
     use linkerd2_proxy_api::{inbound as api, net::InvalidIpNetwork};
 
     #[derive(Debug, thiserror::Error)]
